requirements/requirements.txt
RPi.GPIO
smbus2
ovos_utils
board
neopixel